# app/main.py
# generated by fastapi-codegen:
#   filename:  merged-api-spec.json
#   timestamp: 2025-05-17T13:21:06+00:00

from __future__ import annotations

from typing import Optional, Union

from fastapi import FastAPI, Query

from .models import (
    AirTemperatureGetResponse,
    AirTemperatureGetResponse1,
    AirTemperatureGetResponse2,
    FourDayOutlookGetResponse,
    FourDayOutlookGetResponse1,
    FourDayOutlookGetResponse2,
    TwentyFourHrForecastGetResponse,
    TwentyFourHrForecastGetResponse1,
    TwentyFourHrForecastGetResponse2,
    TwoHrForecastGetResponse,
    TwoHrForecastGetResponse1,
    TwoHrForecastGetResponse2,
    WeatherGetResponse,
    WeatherGetResponse1,
    WeatherGetResponse2,
    WindDirectionGetResponse,
    WindDirectionGetResponse1,
    WindDirectionGetResponse2,
)

app = FastAPI(
    title='Merged API Services',
    description='This is a merged OpenAPI specification containing multiple API services',
    contact={'email': 'feedback@data.gov.sg'},
    version='2025.05.17',
    servers=[{'url': 'https://api-open.data.gov.sg/v2/real-time/api'}],
)


@app.get(
    '/air-temperature',
    response_model=AirTemperatureGetResponse,
    responses={
        '400': {'model': AirTemperatureGetResponse1},
        '404': {'model': AirTemperatureGetResponse2},
    },
)
def get_air_temperature(
    date: Optional[str] = None,
    pagination_token: Optional[str] = Query(None, alias='paginationToken'),
) -> Union[
    AirTemperatureGetResponse, AirTemperatureGetResponse1, AirTemperatureGetResponse2
]:
    """
    Get air temperature readings across Singapore
    """
    pass


@app.get(
    '/four-day-outlook',
    response_model=FourDayOutlookGetResponse,
    responses={
        '400': {'model': FourDayOutlookGetResponse1},
        '404': {'model': FourDayOutlookGetResponse2},
    },
)
def get_four_day_outlook(
    date: Optional[str] = None,
    pagination_token: Optional[str] = Query(None, alias='paginationToken'),
) -> Union[
    FourDayOutlookGetResponse, FourDayOutlookGetResponse1, FourDayOutlookGetResponse2
]:
    """
    Retrieve the latest 4 day weather forecast
    """
    pass


@app.get(
    '/twenty-four-hr-forecast',
    response_model=TwentyFourHrForecastGetResponse,
    responses={
        '400': {'model': TwentyFourHrForecastGetResponse1},
        '404': {'model': TwentyFourHrForecastGetResponse2},
    },
)
def get_twenty_four_hr_forecast(
    date: Optional[str] = None,
    pagination_token: Optional[str] = Query(None, alias='paginationToken'),
) -> Union[
    TwentyFourHrForecastGetResponse,
    TwentyFourHrForecastGetResponse1,
    TwentyFourHrForecastGetResponse2,
]:
    """
    Retrieve the latest 24 hour weather forecast
    """
    pass


@app.get(
    '/two-hr-forecast',
    response_model=TwoHrForecastGetResponse,
    responses={
        '400': {'model': TwoHrForecastGetResponse1},
        '404': {'model': TwoHrForecastGetResponse2},
    },
)
def get_two_hr_forecast(
    date: Optional[str] = None,
    pagination_token: Optional[str] = Query(None, alias='paginationToken'),
) -> Union[
    TwoHrForecastGetResponse, TwoHrForecastGetResponse1, TwoHrForecastGetResponse2
]:
    """
    Retrieve the latest two hour weather forecast
    """
    pass


@app.get(
    '/weather',
    response_model=WeatherGetResponse,
    responses={
        '400': {'model': WeatherGetResponse1},
        '404': {'model': WeatherGetResponse2},
    },
)
def get_weather(
    api: Optional[str] = None,
    date: Optional[str] = None,
    pagination_token: Optional[str] = Query(None, alias='paginationToken'),
) -> Union[WeatherGetResponse, WeatherGetResponse1, WeatherGetResponse2]:
    """
    Retrieve the latest WBGT data for accurate heat stress assessment
    """
    pass


@app.get(
    '/wind-direction',
    response_model=WindDirectionGetResponse,
    responses={
        '400': {'model': WindDirectionGetResponse1},
        '404': {'model': WindDirectionGetResponse2},
    },
)
def get_wind_direction(
    date: Optional[str] = None,
    pagination_token: Optional[str] = Query(None, alias='paginationToken'),
) -> Union[
    WindDirectionGetResponse, WindDirectionGetResponse1, WindDirectionGetResponse2
]:
    """
    Get wind direction readings across Singapore
    """
    pass